fixed tests

+++ src/main/java/com/north/codechallenge/ankitademo/model/Spatial.java
@@ -3,6 +3,7 @@ package com.north.codechallenge.ankitademo.model;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -10,6 +11,6 @@ import lombok.ToString;
 public class Spatial {
 	
 	private String crs;
-	private float[][] bbox;
+	private BigDecimal[][] bbox;
 	
 }
